refactor(macos): remove terminal lifecycle adapter
